% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parts-of-speech.R
\name{parts_of_speech}
\alias{parts_of_speech}
\title{Parts of speech}
\usage{
parts_of_speech("The ripe taste of cheese improves with age.")
}
\arguments{
\item{x}{words}
}
\value{
(list) one vector with words, one with parts of speech
}
\description{
Parts of speech
}
\examples{
parts_of_speech("The ripe taste of cheese improves with age.")
}
