% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swiftspeech-package.R
\docType{package}
\name{swiftspeech}
\alias{swiftspeech}
\alias{swiftspeech-package}
\title{Classify Parts of Speech Using Apple's CoreML and NaturalLanguage Libraries}
\description{
Classify Parts of Speech Using Apple's CoreML and NaturalLanguage Libraries.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://git.rud.is/hrbrmstr/swiftspeech}
  \item Report bugs at \url{https://git.rud.is/hrbrmstr/swiftspeech/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
